% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zips}
\alias{zips}
\title{Zip numbers for Jutland in Denmark}
\format{
A data frame with 53940 rows and 10 variables:

*\link{Zip} Zip code.
\itemize{
\item Area Name of area
}
}
\usage{
zips
}
\description{
A dataset containing the zip numbers for Jutland in Denmark.
}
\keyword{datasets}
